# specs/openapi.app.json
{
  "components": {
    "schemas": {
...
      "ConversationStatus": {
        "enum": [
          "CLOSED",
          "OPEN"
        ],
        "title": "ConversationStatus",
        "type": "string"
      },
      "ConversationView": {
        "properties": {
          "lastAt": {
            "title": "Lastat",
            "type": "string"
          },
          "lastDirection": {
            "$ref": "#/components/schemas/Direction"
          },
          "lastPreview": {
            "title": "Lastpreview",
            "type": "string"
          },
          "lastSenderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsenderid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "peer": {
            "title": "Peer",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ConversationStatus",
            "default": "OPEN"
          },
          "unread": {
            "default": 0,
            "title": "Unread",
            "type": "integer"
          }
        },
        "required": [
          "lastAt",
          "lastDirection",
          "lastPreview",
          "peer",
          "name"
        ],
        "title": "ConversationView",
        "type": "object"
      },
...
      "DemoInboundRequest": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "from": {
            "title": "From",
            "type": "string"
          },
          "previousMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previousmessageid"
          },
          "to": {
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "body",
          "from",
          "to"
        ],
        "title": "DemoInboundRequest",
        "type": "object"
      },
...
      "Direction": {
        "enum": [
          "IN",
          "OUT"
        ],
        "title": "Direction",
        "type": "string"
      },
...
      "Encoding": {
        "enum": [
          "GSM7",
          "UCS2"
        ],
        "title": "Encoding",
        "type": "string"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "MessageRow": {
        "properties": {
          "accountId": {
            "title": "Accountid",
            "type": "string"
          },
          "attemptedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attemptedat"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "campaignId": {
            "title": "Campaignid",
            "type": "string"
          },
          "country": {
            "title": "Country",
            "type": "string"
          },
          "deliveredAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deliveredat"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "encoding": {
            "$ref": "#/components/schemas/Encoding"
          },
          "failureReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failurereason"
          },
          "from": {
            "title": "From",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/Product"
          },
          "mediaKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mediakey"
          },
          "messageId": {
            "title": "Messageid",
            "type": "string"
          },
          "parts": {
            "title": "Parts",
            "type": "integer"
          },
          "priceMicro": {
            "title": "Pricemicro",
            "type": "integer"
          },
          "providerMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providermessageid"
          },
          "queuedAt": {
            "format": "date-time",
            "title": "Queuedat",
            "type": "string"
          },
          "senderId": {
            "default": "",
            "title": "Senderid",
            "type": "string"
          },
          "sentAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentat"
          },
          "status": {
            "$ref": "#/components/schemas/MessageStatus"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "to": {
            "title": "To",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "body",
          "campaignId",
          "country",
          "direction",
          "encoding",
          "from",
          "kind",
          "messageId",
          "parts",
          "priceMicro",
          "queuedAt",
          "status",
          "to",
          "userId",
          "username"
        ],
        "title": "MessageRow",
        "type": "object"
      },
      "MessageStatus": {
        "enum": [
          "DELIVERED",
          "FAILED",
          "QUEUED",
          "RECEIVED",
          "SENT"
        ],
        "title": "MessageStatus",
        "type": "string"
      },
...
      "Page_ConversationView_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConversationView"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[ConversationView]",
        "type": "object"
      },
...
      "Page_MessageRow_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MessageRow"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[MessageRow]",
        "type": "object"
      },
...
      "Product": {
        "enum": [
          "MMS",
          "MMS_AS_LINK",
          "SMS"
        ],
        "title": "Product",
        "type": "string"
      },
...
      "QuickSendResult": {
        "properties": {
          "campaignId": {
            "title": "Campaignid",
            "type": "string"
          },
          "costMicro": {
            "title": "Costmicro",
            "type": "integer"
          },
          "recipients": {
            "title": "Recipients",
            "type": "integer"
          },
          "refused": {
            "items": {
              "$ref": "#/components/schemas/Refusal"
            },
            "title": "Refused",
            "type": "array"
          }
        },
        "required": [
          "campaignId",
          "costMicro",
          "recipients",
          "refused"
        ],
        "title": "QuickSendResult",
        "type": "object"
      },
      "Refusal": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/RefusalReason"
          },
          "to": {
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "message",
          "reason",
          "to"
        ],
        "title": "Refusal",
        "type": "object"
      },
      "RefusalReason": {
        "enum": [
          "COUNTRY_NOT_ENABLED",
          "DAILY_CAP",
          "INSUFFICIENT_BALANCE",
          "NOT_VERIFIED",
          "OPTED_OUT",
          "SENDER_NOT_READY",
          "TRIAL_ENDED"
        ],
        "title": "RefusalReason",
        "type": "string"
      },
      "ReplyRequest": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "senderId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Senderid"
          }
        },
        "required": [
          "body"
        ],
        "title": "ReplyRequest",
        "type": "object"
      },
...
      "ThreadPage": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationView"
          },
          "messages": {
            "$ref": "#/components/schemas/Page_MessageRow_"
          },
          "replyHint": {
            "title": "Replyhint",
            "type": "string"
          }
        },
        "required": [
          "conversation",
          "messages",
          "replyHint"
        ],
        "title": "ThreadPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "txtlocal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/app/conversations": {
      "get": {
        "operationId": "list_conversations_api_app_conversations_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ConversationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ConversationView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "List Conversations"
      }
    },
    "/api/app/conversations/read-all": {
      "post": {
        "operationId": "read_all_api_app_conversations_read_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Read All"
      }
    },
    "/api/app/conversations/{peer}": {
      "get": {
        "operationId": "thread_api_app_conversations__peer__get",
        "parameters": [
          {
            "in": "path",
            "name": "peer",
            "required": true,
            "schema": {
              "title": "Peer",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Thread"
      }
    },
    "/api/app/conversations/{peer}/close": {
      "post": {
        "operationId": "close_api_app_conversations__peer__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "peer",
            "required": true,
            "schema": {
              "title": "Peer",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Close"
      }
    },
    "/api/app/conversations/{peer}/messages": {
      "post": {
        "operationId": "reply_api_app_conversations__peer__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "peer",
            "required": true,
            "schema": {
              "title": "Peer",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickSendResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Reply"
      }
    },
    "/api/app/conversations/{peer}/read": {
      "post": {
        "operationId": "mark_read_api_app_conversations__peer__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "peer",
            "required": true,
            "schema": {
              "title": "Peer",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Mark Read"
      }
    },
    "/api/app/conversations/{peer}/reopen": {
      "post": {
        "operationId": "reopen_api_app_conversations__peer__reopen_post",
        "parameters": [
          {
            "in": "path",
            "name": "peer",
            "required": true,
            "schema": {
              "title": "Peer",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Reopen"
      }
    },
...
    "/api/app/demo/inbound": {
      "post": {
        "operationId": "demo_inbound_api_app_demo_inbound_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInboundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused"
          }
        },
        "summary": "Demo Inbound"
      }
    },
...
  }
}